% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway-inators.R
\name{GWsFavoriteKEGGPathways}
\alias{GWsFavoriteKEGGPathways}
\title{GWsFavoriteKEGGPathways
A function to return a vector of KEGG pathways that are commonly studied in GW's research.}
\usage{
GWsFavoriteKEGGPathways(KEGG_IDs = FALSE)
}
\arguments{
\item{KEGG_IDs}{Logical, if TRUE returns KEGG pathway IDs, if FALSE returns pathway names. Default is FALSE.}
}
\value{
A vector of KEGG pathway IDs or names.
}
\description{
GWsFavoriteKEGGPathways
A function to return a vector of KEGG pathways that are commonly studied in GW's research.
}
